Need to know how to use actions in redux before making autoregister toggle
