--- src/settings/components/AutoRegister/AutoRegisterContainer.js
@@ -1,20 +1,0 @@
-import { connect } from 'react-redux';
-import AutoRegister from './AutoRegister';
-import dialBackendApi from '../../../services/api';
-
-function mapStateToProps({ user }) {
-  return {
-    autoregister: user.autoregister
-  };
-}
-
-function mapDispatchToProps() {
-  return ({
-    setAutoRegister: dialBackendApi().setAutoRegister
-  });
-}
-
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps
-)(AutoRegister);
