+++ src/settings/screens/AutoRegister/AutoRegister.js
@@ -1,10 +1,11 @@
 import React from 'react';
 import { Text, Card } from 'react-native-elements';
-import { RadioButton } from 'react-native-paper';
 import { ListItem, Button, Icon } from 'react-native-elements';
 
-const AutoRegister = ({ autoregister, setAutoRegister }) => {
-    console.log(autoregister);
+const AutoRegister = (props) => {
+  console.log("props passed to autoregister 0000000000000000", props);
+  const { autoRegister, toggle_autoregister } = props;
+  console.log(autoRegister, toggle_autoregister);
   return (
     <Card>
       <Text style={{ fontWeight: 'bold' }}>
@@ -12,7 +13,10 @@ const AutoRegister = ({ autoregister, setAutoRegister }) => {
       </Text>
       <ListItem
         title="Auto-register on start"
-        switch={{ onValueChange: ()=>setAutoRegister(!autoregister), value: (autoregister) ? true : false }}
+        switch={{
+          onValueChange: () => toggle_autoregister(!autoRegister),
+          value: autoRegister
+        }}
       />
     </Card>
   );
